# Bw6 miller loop optimization