--- a/ec/src/models/bw6/g2.rs
+++ b/ec/src/models/bw6/g2.rs
@@ -84,26 +84,21 @@ impl<P: BW6Config> From<G2Affine<P>> for G2Prepared<P> {
         }
         ell_coeffs_1.push(r.clone().add_in_place(&q));
 
-        // f_{u^3-u^2-u,Q}(P)
+        // f_{u^2-u-1,Q}(P)
         let qu: G2Affine<P> = r.into();
         assert_eq!(qu, q.mul_bigint(&P::ATE_LOOP_COUNT_1));
         assert_eq!(qu, q.mul_bigint(&P::X));
 
         let mut ell_coeffs_2 = vec![];
-        let mut r = G2HomProjective::<P> {
-            x: q.x,
-            y: q.y,
-            z: P::Fp::one(),
-        };
 
-        let neg_q = -q;
+        let neg_qu = -qu;
 
         for bit in P::ATE_LOOP_COUNT_2.iter().rev().skip(1) {
             ell_coeffs_2.push(r.double_in_place());
 
             match bit {
-                1 => ell_coeffs_2.push(r.add_in_place(&q)),
-                -1 => ell_coeffs_2.push(r.add_in_place(&neg_q)),
+                1 => ell_coeffs_2.push(r.add_in_place(&qu)),
+                -1 => ell_coeffs_2.push(r.add_in_place(&neg_qu)),
                 _ => continue,
             }
         }