--- a/ec/src/models/bw6/mod.rs
+++ b/ec/src/models/bw6/mod.rs
@@ -95,7 +95,7 @@ pub trait BW6Config: 'static + Eq + Sized {
 
         let f_u_inv = f_u.cyclotomic_inverse().unwrap();
 
-        // TODO: figure
+        // TODO: is it enough to get the inverse of f_1, or does f_u also need to get inverted?
         if Self::ATE_LOOP_COUNT_1_IS_NEGATIVE {
             f_u.cyclotomic_inverse_in_place();
         }