<!-- < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < ☺
v                               ✰  Thanks for creating a PR! ✰
v    Before hitting that submit button please review the checkboxes.
v    If a checkbox is n/a - please still include it but + a little note why
☺ > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > >  -->

## Description

As per Algorithm 5 in https://eprint.iacr.org/2020/351.pdf.

TODO: check whether the inverse should apply to `f_u` or `f_1` in case the loop count is negative.
@yelhousni maybe you could help?

Benchmarks:
Before:
```
Pairing for BW6_761/G2 Preparation for BW6_761
                        time:   [732.74 µs 734.78 µs 737.28 µs]
Pairing for BW6_761/Miller Loop for BW6_761
                        time:   [2.5692 ms 2.5765 ms 2.5844 ms]
Pairing for BW6_761/Full Pairing for BW6_761
                        time:   [6.6391 ms 6.6557 ms 6.6730 ms]
```
After:
```
Pairing for BW6_761/G2 Preparation for BW6_761
                        time:   [591.47 µs 592.55 µs 593.67 µs]
Pairing for BW6_761/Miller Loop for BW6_761
                        time:   [2.0311 ms 2.0358 ms 2.0406 ms]
Pairing for BW6_761/Full Pairing for BW6_761
                        time:   [5.9286 ms 5.9474 ms 5.9669 ms]
```
full pairing is ~11% faster.

closes: #616 

---

Before we can merge this PR, please make sure that all the following items have been
checked off. If any of the checklist items are not applicable, please leave them but
write a little note why.

- [x] Targeted PR against correct branch (master)
- [x] Linked to GitHub issue with discussion and accepted design OR have an explanation in the PR that describes this work.
- [x] Wrote unit tests
- [x] Updated relevant documentation in the code
- [ ] Added a relevant changelog entry to the `Pending` section in `CHANGELOG.md`
- [x] Re-reviewed `Files changed` in the GitHub PR explorer