@mmagician 2023-03-11
We can probably remove the check for whether 2nd Miller loop is negative. The parameter X^2 - X - 1 > 0, unless `P::X=1` I don't know enough about curve construction, but I guess this will never happen.
Also we could basically reuse `P::X`, and its sign, to determine the value and sign of the first loop count. Maybe to even find the 2-NAF of the 2nd count with a `const fn`. What's standing in the way is that X is currently defined as Bigint. Maybe in a next PR I can refactor the BW6 parameters, ideally removing these 4 parameters that are already known/can be computed at compile.