@mmagician 2023-09-01
@Pratyush The parameters for loop counts changed, so the tests will fail against the current master of arkworks/curves. I've updated both sides today though, and they are all green - I can add a temp patch to Cargo here to point to the compatible branch in curves.
However, one thing to reconsider (as pointed out to me by @swasilyev at some point) is that even though this algorithm is better for a single pairing, it seems that with a multipairing of 3 the methods are equivalent, and likely worsening the performance for 4+, see the last table in https://hackmd.io/@gnark/BW6-761-changes.
There seems to be one method (called 6' in that hackmd) which is **both** backwards compatible, faster for single and for multi-pairings. I could implement that, but won't be able to do it very soon. WDYT, how likely are we going to do multi-pairings of size 4+ vs. single pairings?