@Pratyush 2023-09-02
I think multi-pairings often with some frequency, especially for batch-verification.