@mmagician 2023-09-02
I ran some benchmarks, and it seems that the new version from this PR outperforms the current master, even for multi-pairings.

Some numbers for BW6-767:
```
optimized:

Pairing for BW6_767/Multi Pairing for BW6_767 with 1 pairs
                        time:   [3.9839 ms 4.0180 ms 4.0593 ms]
Pairing for BW6_767/Multi Pairing for BW6_767 with 5 pairs
                        time:   [8.7305 ms 8.7505 ms 8.7706 ms]
Pairing for BW6_767/Multi Pairing for BW6_767 with 10 pairs
                        time:   [14.508 ms 14.563 ms 14.617 ms]

master:

Pairing for BW6_767/Multi Pairing for BW6_767 with 1 pairs
                        time:   [4.4910 ms 4.5152 ms 4.5490 ms]
Pairing for BW6_767/Multi Pairing for BW6_767 with 5 pairs
                        time:   [11.000 ms 11.046 ms 11.092 ms]
Pairing for BW6_767/Multi Pairing for BW6_767 with 10 pairs
                        time:   [18.986 ms 19.064 ms 19.145 ms]
```

Where the improvement % actually **increases** with number of pairs!

Similar results for BW6-761:
```
optimized:

Pairing for BW6_761/Multi Pairing for BW6_761 with 1 pairs
                        time:   [3.5187 ms 3.5259 ms 3.5333 ms]
Pairing for BW6_761/Multi Pairing for BW6_761 with 5 pairs
                        time:   [8.1297 ms 8.2073 ms 8.3008 ms]
Pairing for BW6_761/Multi Pairing for BW6_761 with 10 pairs
                        time:   [13.810 ms 13.860 ms 13.911 ms]

master:

Pairing for BW6_761/Multi Pairing for BW6_761 with 1 pairs
                        time:   [3.9132 ms 3.9271 ms 3.9413 ms]
Pairing for BW6_761/Multi Pairing for BW6_761 with 5 pairs
                        time:   [9.9494 ms 9.9683 ms 9.9876 ms]
Pairing for BW6_761/Multi Pairing for BW6_761 with 10 pairs
                        time:   [17.237 ms 17.273 ms 17.309 ms]
```

I'm not quite sure what to make of this given the notes in https://hackmd.io/@gnark/BW6-761-changes. I suppose it comes down to the implementation differences across the gnark & arkworks at some different level in the stack.

In any case, I think the numbers suggest we can safely call this PR an improved version.